Refactor Project Structure
To have a src folder and be able to have a test folder too
