+++ src/karthuria/model/school.py
